Move api to bot process
