+++ bot/commands/ping.ts
@@ -1,6 +1,7 @@
 import { SlashCommand, SlashCreator, CommandContext } from "slash-create";
+import { BotClient } from "../types";
 
-export default class Ping extends SlashCommand {
+export default class Ping extends SlashCommand<BotClient> {
 	constructor(creator: SlashCreator) {
 		super(creator, {
 			name: "ping",
