+++ bot/prisma/schema.prisma
@@ -1,0 +1,14 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "sqlite"
+  url      = env("DATABASE_URL")
+}
+
+model Room {
+  id            Int    @id @default(autoincrement())
+  room_id       String @unique
+  hb_session_id String @unique
+}
