+++ bot/types.d.ts
@@ -1,0 +1,6 @@
+import { PrismaClient } from "@prisma/client";
+import { Client } from "discord.js";
+
+export interface BotClient extends Client {
+	db: PrismaClient;
+}
