+++ package.json
@@ -8,7 +8,6 @@
 		"start": "pm2 start ./pm2.config.js",
 		"bot": "pnpm run build:bot && pnpm run launch:bot",
 		"client": "pnpm run build:client && pnpm run launch:client",
-		"api": "pnpm run build:api && pnpm run launch:api",
 		"listcommands": "slash-up local",
 		"build:bot": "tsc --project tsconfig.bot.json",
 		"launch:bot": "cd dist/bot && node index.js",
@@ -17,12 +16,11 @@
 		"launch:client": "vite",
 		"build:client": "vite build",
 		"preview:client": "vite preview",
-		"build:api": "tsc --project tsconfig.api.json",
-		"launch:api": "cd dist/api && node index.js",
-		"envtypes": "pnpm exec dotenv-types-generator --file ./.env --mergeTypes true"
+		"envtypes": "pnpm exec dotenv-types-generator --file ./.env"
 	},
 	"dependencies": {
 		"@hyperbeam/iframe": "^0.0.10",
+		"@prisma/client": "^4.0.0",
 		"@types/better-sqlite3": "^7.5.0",
 		"@vitejs/plugin-react": "^1.3.2",
 		"better-sqlite3": "^7.5.3",
@@ -48,8 +46,12 @@
 		"eslint-plugin-jsx-a11y": "^6.6.0",
 		"eslint-plugin-react": "^7.30.1",
 		"eslint-plugin-simple-import-sort": "^7.0.0",
+		"prisma": "^4.0.0",
 		"slash-up": "^1.2.0",
 		"typescript": "^4.7.4",
 		"vite-plugin-mkcert": "^1.7.2"
+	},
+	"prisma": {
+		"schema": "bot/prisma/schema.prisma"
 	}
 }
