--- tsconfig.api.json
@@ -1,9 +1,0 @@
-{
-	"compilerOptions": {
-		"outDir": "dist/api",
-		"esModuleInterop": true,
-	},
-	"include": [
-		"api",
-	],
-}
